+++ src/main/java/eu/gir/gircredstone/init/GIRCInit.java
@@ -1,7 +1,6 @@
 package eu.gir.gircredstone.init;
 
-import java.util.ArrayList;
-import java.util.List;
+import java.util.function.Supplier;
 
 import eu.gir.gircredstone.GIRCRedstoneMain;
 import eu.gir.gircredstone.block.BlockRedstoneAcceptor;
@@ -14,44 +13,38 @@ import net.minecraft.world.item.Item;
 import net.minecraft.world.item.Item.Properties;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.entity.BlockEntityType;
-import net.minecraftforge.event.RegistryEvent;
-import net.minecraftforge.eventbus.api.SubscribeEvent;
-import net.minecraftforge.fml.common.Mod;
-import net.minecraftforge.fml.common.Mod.EventBusSubscriber.Bus;
-import net.minecraftforge.registries.IForgeRegistry;
+import net.minecraft.world.level.block.state.BlockBehaviour;
+import net.minecraft.world.level.material.Material;
+import net.minecraftforge.eventbus.api.IEventBus;
+import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
+import net.minecraftforge.registries.DeferredRegister;
+import net.minecraftforge.registries.ForgeRegistries;
+import net.minecraftforge.registries.RegistryObject;
 
-@Mod.EventBusSubscriber(bus = Bus.MOD, modid = GIRCRedstoneMain.MODID)
 public class GIRCInit {
 	
-	public static Block RS_ACCEPTOR;
-	public static Block RS_EMITTER;
+	public static final DeferredRegister<Item> ITEM_REGISTRY = DeferredRegister.create(ForgeRegistries.ITEMS, GIRCRedstoneMain.MODID);
+	public static final DeferredRegister<Block> BLOCK_REGISTRY = DeferredRegister.create(ForgeRegistries.BLOCKS, GIRCRedstoneMain.MODID);
+	public static final DeferredRegister<BlockEntityType<?>> TILEENTITY_REGISTRY = DeferredRegister.create(ForgeRegistries.BLOCK_ENTITIES, GIRCRedstoneMain.MODID);
 	
-	public static Item RS_LINKER;
+	public static RegistryObject<Block> RS_ACCEPTOR = internalRegisterBlock("acceptor", () -> new BlockRedstoneAcceptor(BlockBehaviour.Properties.of(Material.METAL).strength(1.5f, 6.0f).requiresCorrectToolForDrops()));
+	public static RegistryObject<Block> RS_EMITTER = internalRegisterBlock("emitter", () -> new BlockRedstoneEmitter(BlockBehaviour.Properties.of(Material.METAL).strength(1.5f, 6.0f).requiresCorrectToolForDrops()));
 	
-	private static final List<Block> blocks = new ArrayList<>();
+	public static RegistryObject<Item> RS_LINKER = ITEM_REGISTRY.register("linker", () -> new Linkingtool(new Properties().tab(CreativeModeTab.TAB_REDSTONE)));
 	
-	@SubscribeEvent
-	public static void registerBlock(final RegistryEvent.Register<Block> event) {
-		final IForgeRegistry<Block> registry = event.getRegistry();
-		blocks.add(RS_ACCEPTOR = new BlockRedstoneAcceptor());
-		blocks.add(RS_EMITTER = new BlockRedstoneEmitter());
-		blocks.forEach(registry::register);
+	public static RegistryObject<BlockEntityType<?>> EMITER_TILE = TILEENTITY_REGISTRY.register("emitter", () -> BlockEntityType.Builder.of(TileRedstoneEmitter::new, RS_EMITTER.get()).build(null));
+	
+	private static RegistryObject<Block> internalRegisterBlock(final String name, final Supplier<Block> sup) {
+		final RegistryObject<Block> registerObject = BLOCK_REGISTRY.register(name, sup);
+		ITEM_REGISTRY.register(name, () -> new BlockItem(registerObject.get(), new Properties().tab(CreativeModeTab.TAB_REDSTONE)));
+		return registerObject;
 	}
 	
-	@SubscribeEvent
-	public static void registerItem(final RegistryEvent.Register<Item> event) {
-		final IForgeRegistry<Item> registry = event.getRegistry();
-		blocks.forEach(block -> registry.register(new BlockItem(block, new Properties().tab(CreativeModeTab.TAB_REDSTONE)).setRegistryName(block.getRegistryName())));
-		registry.register(RS_LINKER = new Linkingtool());
-	}
-	
-	public static BlockEntityType<?> EMITER_TILE;
-	
-	@SubscribeEvent
-	public static void registerTE(final RegistryEvent.Register<BlockEntityType<?>> evt) {
-		EMITER_TILE = BlockEntityType.Builder.of(TileRedstoneEmitter::new, RS_EMITTER).build(null);
-		EMITER_TILE.setRegistryName(GIRCRedstoneMain.MODID, "emitter");
-		evt.getRegistry().register(EMITER_TILE);
+	public static void init() {
+		final IEventBus bus = FMLJavaModLoadingContext.get().getModEventBus();
+		ITEM_REGISTRY.register(bus);
+		BLOCK_REGISTRY.register(bus);
+		TILEENTITY_REGISTRY.register(bus);
 	}
 	
 }