+++ changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## [Update/Fix]
+
+* feat: Added block tags and loot tables (blocks drop items when they are broken)
 ## [Internal]
 
 * Internal forge update